requirements.txt
bs4~=0.0.1
Flask~=2.2.2
pymongo~=4.3.3
waitress~=2.1.2
requests~=2.28.1
jsonobject~=2.1.0
beautifulsoup4~=4.11.1
strawberry-graphql~=0.142.2
